requirements.txt
faiss-cpu
pdfplumber
PyMuPDF
nltk
rank_bm25
gradio==3.50.2
matplotlib
numpy
pandas
pydantic==1.10.13
tqdm
sentence-transformers
groq
google-generativeai
python-dotenv
lime
scikit-learn
accelerate
opencv-python-headless